requirements.txt
streamlit>=1.36
pypdf>=5.0
groq>=0.4.0
python-dotenv>=1.0.0
